p2p: Combining client and node.
